refining react structure
