+++ package.json
@@ -9,6 +9,7 @@
     "react-dom": "^16.3.2",
     "react-google-charts": "^3.0.10",
     "react-google-maps": "^9.4.5",
+    "react-loadable": "^5.5.0",
     "react-paginate": "^5.2.3",
     "react-redux": "^6.0.0",
     "react-router": "^3.2.1",
