--- src/components/App.js
@@ -1,17 +1,0 @@
-import React, { Component } from 'react';
-import logo from '../logo.svg';
-import '../App.css';
-
-import Layout from '../layouts/index';
-
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <Layout />
-      </div>
-    );
-  }
-}
-
-export default App;
