--- src/index.css
@@ -1,18 +1,0 @@
-body {
-  margin-left: 20%;
-  padding: 0;
-  font-family: sans-serif;
-}
-
-/* Sandy */
-.pad-200 {
-  padding-right: 230px !important;
-}
-
-.card-block {
-  padding-right: 200px;
-}
-
-.left-al {
-  margin-left: -750px
-}
