+++ src/styles/index.css
@@ -1,0 +1,6 @@
+.routable{
+  padding: 100px 100px 0px 0px;
+}
+.heading{
+text-align: center;
+}
